Init first release
