+++ CHANGELOG.md
@@ -4,6 +4,6 @@ This changelog follows [the Keep a Changelog standard](https://keepachangelog.co
 
 ## 1.0.0 (2022-02-25)
 
-Initial release. Synced with the latest commit [Shopify Icons - v4.17.0](https://www.npmjs.com/package/@shopify/polaris-icons/v/4.17.0)
+Initial release. Synced with the latest commit [Polaris Icons - v4.17.0](https://www.npmjs.com/package/@shopify/polaris-icons/v/4.17.0)
