+++ config/blade-polaris-icons.php
@@ -13,7 +13,7 @@ return [
     |
     */
 
-    'prefix' => 'shi',
+    'prefix' => 'polaris',
 
     /*
     |-----------------------------------------------------------------
